+++ packages/cli/commands/project.js
@@ -6,6 +6,7 @@ const {
 const deploy = require('./project/deploy');
 const create = require('./project/create');
 const upload = require('./project/upload');
+const logs = require('./project/logs');
 
 exports.command = 'project';
 exports.describe = false; //'Commands for working with projects';
@@ -19,6 +20,7 @@ exports.builder = yargs => {
   yargs.command(deploy).demandCommand(1, '');
   yargs.command(create).demandCommand(0, '');
   yargs.command(upload).demandCommand(0, '');
+  yargs.command(logs).demandCommand(1, '');
 
   return yargs;
 };