+++ packages/cli/lib/projects.js
@@ -12,18 +12,43 @@ const { getHubSpotWebsiteOrigin } = require('@hubspot/cli-lib/lib/urls');
 const {
   ENVIRONMENTS,
   POLLING_DELAY,
-  PROJECT_BUILD_STATUS,
-  PROJECT_BUILD_STATUS_TEXT,
-  PROJECT_DEPLOY_STATUS,
-  PROJECT_DEPLOY_STATUS_TEXT,
   PROJECT_TEMPLATE_REPO,
+  PROJECT_OVERALL_STATUS,
+  PROJECT_TEXT,
 } = require('@hubspot/cli-lib/lib/constants');
-const { getBuildStatus, getDeployStatus } = require('@hubspot/cli-lib/api/dfs');
+const {
+  getBuildStatus,
+  getDeployStatus,
+  fetchProject,
+} = require('@hubspot/cli-lib/api/dfs');
+const {
+  logApiErrorInstance,
+  ApiErrorContext,
+} = require('@hubspot/cli-lib/errorHandlers');
 
-const isBuildComplete = build => {
+const PROJECT_STRINGS = {
+  BUILD: {
+    INITIALIZE: (name, numOfComponents) =>
+      `Building ${chalk.bold(
+        name
+      )}\n\nFound ${numOfComponents} components in this project ...\n`,
+    SUCCESS: name => `Built ${chalk.bold(name)}`,
+    FAIL: name => `Failed to build ${chalk.bold(name)}`,
+  },
+  DEPLOY: {
+    INITIALIZE: (name, numOfComponents) =>
+      `Deploying ${chalk.bold(
+        name
+      )}\n\nFound ${numOfComponents} components in this project ...\n`,
+    SUCCESS: name => `Deployed ${chalk.bold(name)}`,
+    FAIL: name => `Failed to deploy ${chalk.bold(name)}`,
+  },
+};
+
+const isTaskComplete = task => {
   return (
-    build.status === PROJECT_BUILD_STATUS.SUCCESS ||
-    build.status === PROJECT_BUILD_STATUS.FAILURE
+    task.status === PROJECT_OVERALL_STATUS.SUCCESS ||
+    task.status === PROJECT_OVERALL_STATUS.FAILURE
   );
 };
 
@@ -125,7 +150,7 @@ const createProjectConfig = async (projectPath, projectName) => {
   return projectConfig;
 };
 
-const validateProjectConfig = projectConfig => {
+const validateProjectConfig = (projectConfig, projectDir) => {
   if (!projectConfig) {
     logger.error(
       `Project config not found. Try running 'hs project init' first.`
@@ -140,11 +165,42 @@ const validateProjectConfig = projectConfig => {
     process.exit(1);
   }
 
-  if (!fs.existsSync(projectConfig.srcDir)) {
+  if (!fs.existsSync(path.resolve(projectDir, projectConfig.srcDir))) {
     logger.error(
       `Project source directory '${projectConfig.srcDir}' does not exist.`
     );
     process.exit(1);
+  }
+};
+
+const ensureProjectExists = async (accountId, projectName) => {
+  try {
+    await fetchProject(accountId, projectName);
+  } catch (err) {
+    if (err.statusCode === 404) {
+      const { shouldCreateProject } = await prompt([
+        {
+          name: 'shouldCreateProject',
+          message: `The project ${projectName} does not exist in ${accountId}. Would you like to create it?`,
+          type: 'confirm',
+        },
+      ]);
+
+      if (shouldCreateProject) {
+        try {
+          return createProject(accountId, projectName);
+        } catch (err) {
+          return logApiErrorInstance(err, new ApiErrorContext({ accountId }));
+        }
+      } else {
+        return logger.log(
+          `Your project ${chalk.bold(
+            projectName
+          )} could not be found in ${chalk.bold(accountId)}.`
+        );
+      }
+    }
+    logApiErrorInstance(err, new ApiErrorContext({ accountId }));
   }
 };
 
@@ -178,183 +234,107 @@ const showWelcomeMessage = () => {
   logger.log('-------------------------------------------------------------');
 };
 
-const pollBuildStatus = async (accountId, name, buildId) => {
-  const buildStatus = await getBuildStatus(accountId, name, buildId);
-  const spinnies = new Spinnies();
-
-  logger.log();
-  logger.log(`Building ${chalk.bold(name)}`);
-  logger.log();
-  logger.log(`Found ${buildStatus.subbuildStatuses.length} deployables ...`);
-  logger.log();
-
-  for (let subBuild of buildStatus.subbuildStatuses) {
-    spinnies.add(subBuild.buildName, {
-      text: `${chalk.bold(subBuild.buildName)} #${buildId} ${
-        PROJECT_BUILD_STATUS_TEXT[PROJECT_BUILD_STATUS.ENQUEUED]
-      }`,
-    });
+const makeGetTaskStatus = taskType => {
+  let statusFn, statusText, statusStrings;
+  switch (taskType) {
+    case 'build':
+      statusFn = getBuildStatus;
+      statusText = PROJECT_TEXT.BUILD;
+      statusStrings = PROJECT_STRINGS.BUILD;
+      break;
+    case 'deploy':
+      statusFn = getDeployStatus;
+      statusText = PROJECT_TEXT.DEPLOY;
+      statusStrings = PROJECT_STRINGS.DEPLOY;
+      break;
+    default:
+      logger.error(`Cannot get status for task type ${taskType}`);
   }
 
-  return new Promise((resolve, reject) => {
-    const pollInterval = setInterval(async () => {
-      const buildStatus = await getBuildStatus(accountId, name, buildId).catch(
-        reject
-      );
-      const { status, subbuildStatuses } = buildStatus;
-
-      if (spinnies.hasActiveSpinners()) {
-        subbuildStatuses.forEach(subBuild => {
-          if (!spinnies.pick(subBuild.buildName)) {
-            return;
-          }
-
-          const updatedText = `${chalk.bold(subBuild.buildName)} #${buildId} ${
-            PROJECT_BUILD_STATUS_TEXT[subBuild.status]
-          }`;
-
-          switch (subBuild.status) {
-            case PROJECT_BUILD_STATUS.SUCCESS:
-              spinnies.succeed(subBuild.buildName, {
-                text: updatedText,
-              });
-              break;
-            case PROJECT_BUILD_STATUS.FAILURE:
-              spinnies.fail(subBuild.buildName, {
-                text: updatedText,
-              });
-              break;
-            default:
-              spinnies.update(subBuild.buildName, {
-                text: updatedText,
-              });
-              break;
-          }
-        });
-      }
-
-      if (isBuildComplete(buildStatus)) {
-        clearInterval(pollInterval);
-
-        if (status === PROJECT_BUILD_STATUS.SUCCESS) {
-          logger.success(
-            `Your project ${chalk.bold(name)} ${
-              PROJECT_BUILD_STATUS_TEXT[status]
-            }.`
-          );
-        } else if (status === PROJECT_BUILD_STATUS.FAILURE) {
-          logger.error(
-            `Your project ${chalk.bold(name)} ${
-              PROJECT_BUILD_STATUS_TEXT[status]
-            }.`
-          );
-          subbuildStatuses.forEach(subBuild => {
-            if (subBuild.status === PROJECT_BUILD_STATUS.FAILURE) {
-              logger.error(
-                `${chalk.bold(subBuild.buildName)} failed to build. ${
-                  subBuild.errorMessage
-                }.`
-              );
-            }
-          });
-        }
-        resolve(buildStatus);
-      }
-    }, POLLING_DELAY);
-  });
-};
-
-const pollDeployStatus = async (accountId, name, deployId, deployedBuildId) => {
-  const deployStatus = await getDeployStatus(accountId, name, deployId);
-  const spinnies = new Spinnies();
-
-  logger.log();
-  logger.log(`Deploying ${chalk.bold(name)}`);
-  logger.log();
-  logger.log(
-    `Found ${deployStatus.subdeployStatuses.length} sub-build deploys ...`
-  );
-  logger.log();
-
-  for (let subdeploy of deployStatus.subdeployStatuses) {
-    spinnies.add(subdeploy.deployName, {
-      text: `${chalk.bold(subdeploy.deployName)} #${deployedBuildId} ${
-        PROJECT_DEPLOY_STATUS_TEXT[PROJECT_DEPLOY_STATUS.ENQUEUED]
-      }`,
+  return async (accountId, taskName, taskId) => {
+    const spinnies = new Spinnies({
+      succeedColor: 'white',
+      failColor: 'white',
     });
-  }
 
-  return new Promise((resolve, reject) => {
-    const pollInterval = setInterval(async () => {
-      const deployStatus = await getDeployStatus(
-        accountId,
-        name,
-        deployId
-      ).catch(reject);
+    spinnies.add('overallTaskStatus', { text: 'Beginning' });
 
-      const { status, subdeployStatuses } = deployStatus;
+    const initialTaskStatus = await statusFn(accountId, taskName, taskId);
 
-      if (spinnies.hasActiveSpinners()) {
-        subdeployStatuses.forEach(subdeploy => {
-          if (!spinnies.pick(subdeploy.deployName)) {
-            return;
-          }
+    spinnies.update('overallTaskStatus', {
+      text: statusStrings.INITIALIZE(
+        taskName,
+        initialTaskStatus[statusText.SUBTASK_KEY].length
+      ),
+    });
 
-          const updatedText = `${chalk.bold(
-            subdeploy.deployName
-          )} #${deployedBuildId} ${
-            PROJECT_DEPLOY_STATUS_TEXT[subdeploy.status]
-          }`;
+    for (let subTask of initialTaskStatus[statusText.SUBTASK_KEY]) {
+      spinnies.add(subTask[statusText.SUBTASK_NAME_KEY], {
+        text: `${chalk.bold(subTask[statusText.SUBTASK_NAME_KEY])} #${taskId} ${
+          statusText.STATUS_TEXT[statusText.STATES.ENQUEUED]
+        }\n`,
+      });
+    }
 
-          switch (subdeploy.status) {
-            case PROJECT_DEPLOY_STATUS.SUCCESS:
-              spinnies.succeed(subdeploy.deployName, {
-                text: updatedText,
-              });
-              break;
-            case PROJECT_DEPLOY_STATUS.FAILURE:
-              spinnies.fail(subdeploy.deployName, {
-                text: updatedText,
-              });
-              break;
-            default:
-              spinnies.update(subdeploy.deployName, {
-                text: updatedText,
-              });
-              break;
-          }
-        });
-      }
+    return new Promise((resolve, reject) => {
+      const pollInterval = setInterval(async () => {
+        const taskStatus = await statusFn(accountId, taskName, taskId).catch(
+          reject
+        );
 
-      if (isBuildComplete(deployStatus)) {
-        clearInterval(pollInterval);
+        const { status, [statusText.SUBTASK_KEY]: subTaskStatus } = taskStatus;
 
-        if (status === PROJECT_DEPLOY_STATUS.SUCCESS) {
-          logger.success(
-            `Your project ${chalk.bold(name)} ${
-              PROJECT_DEPLOY_STATUS_TEXT[status]
-            }.`
-          );
-        } else if (status === PROJECT_DEPLOY_STATUS.FAILURE) {
-          logger.error(
-            `Your project ${chalk.bold(name)} ${
-              PROJECT_DEPLOY_STATUS_TEXT[status]
-            }.`
-          );
-          subdeployStatuses.forEach(subdeploy => {
-            if (subdeploy.status === PROJECT_DEPLOY_STATUS.FAILURE) {
-              logger.error(
-                `${chalk.bold(subdeploy.deployName)} failed to build. ${
-                  subdeploy.errorMessage
-                }.`
-              );
+        if (spinnies.hasActiveSpinners()) {
+          subTaskStatus.forEach(subTask => {
+            if (!spinnies.pick(subTask[statusText.SUBTASK_NAME_KEY])) {
+              return;
+            }
+
+            const updatedText = `${chalk.bold(
+              subTask[statusText.SUBTASK_NAME_KEY]
+            )} #${taskId} ${statusText.STATUS_TEXT[subTask.status]}\n`;
+
+            switch (subTask.status) {
+              case statusText.STATES.SUCCESS:
+                spinnies.succeed(subTask[statusText.SUBTASK_NAME_KEY], {
+                  text: updatedText,
+                });
+                break;
+              case statusText.STATES.FAILURE:
+                spinnies.fail(subTask.buildName, {
+                  text: updatedText,
+                });
+                break;
+              default:
+                spinnies.update(subTask.buildName, {
+                  text: updatedText,
+                });
+                break;
             }
           });
+
+          if (isTaskComplete(taskStatus)) {
+            subTaskStatus.forEach(subBuild => {
+              spinnies.remove(subBuild[statusText.SUBTASK_NAME_KEY]);
+            });
+
+            if (status === statusText.STATES.SUCCESS) {
+              spinnies.succeed('overallTaskStatus', {
+                text: statusStrings.SUCCESS(taskName),
+              });
+            } else if (status === statusText.STATES.FAILURE) {
+              spinnies.fail('overallTaskStatus', {
+                text: statusStrings.FAIL(taskName),
+              });
+            }
+
+            clearInterval(pollInterval);
+            resolve(taskStatus);
+          }
         }
-        resolve(deployStatus);
-      }
-    }, POLLING_DELAY);
-  });
+      }, POLLING_DELAY);
+    });
+  };
 };
 
 module.exports = {
@@ -363,7 +343,8 @@ module.exports = {
   createProjectConfig,
   validateProjectConfig,
   showWelcomeMessage,
-  pollBuildStatus,
-  pollDeployStatus,
   getProjectDetailUrl,
+  pollBuildStatus: makeGetTaskStatus('build'),
+  pollDeployStatus: makeGetTaskStatus('deploy'),
+  ensureProjectExists,
 };