+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 3.0.9
+
+- Fixed issue in theme marketplace-validate ([#556](https://github.com/HubSpot/hubspot-cli/pull/556))
+- Encode project name in request URLs ([#566](https://github.com/HubSpot/hubspot-cli/pull/566))
+- Bump tmpl from 1.0.4 to 1.0.5 ([#560](https://github.com/HubSpot/hubspot-cli/pull/560))
+- Support section validation in theme marketplace-validate ([#546](https://github.com/HubSpot/hubspot-cli/pull/546))
+- Fix project build status bug on build failure ([#562](https://github.com/HubSpot/hubspot-cli/pull/562))
+- Update project upload command to handle initialization ([#564](https://github.com/HubSpot/hubspot-cli/pull/564))
+- Fix template util bug ([#556](https://github.com/HubSpot/hubspot-cli/pull/556))
+- Update project deploy command ([#561](https://github.com/HubSpot/hubspot-cli/pull/561))
+
 ## 3.0.8
 
 - Marketplace theme validation has been moved out of beta ([#553](https://github.com/HubSpot/hubspot-cli/pull/553))